telegram-notes/app/requirements.txt
flask
python-dotenv
pyTelegramBotAPI
pymysql
loguru
openai>=0.27.0